Fix: Move state to hook
