+++ src/components/@helpers/dropdown/dummy_api.ts
@@ -1,0 +1,17 @@
+export const api_dummy_response = [
+  {
+    id: '3r56',
+    label: 'First',
+    value: '1st'
+  },
+  {
+    id: '7lh0',
+    label: 'Second',
+    value: '2nd'
+  },
+  {
+    id: 'r89y',
+    label: 'Third',
+    value: '3rd'
+  },
+];
